#3 Create Staging Config and add Configuration to enable DB Seeder
Created Stating Configuration to be used in the testing environment.
This file can be enabled by setting the spring profile to "STAGING",
which should be achievable by setting the environment var
SPRING_PROFILES_ACTIVE to STAGING.

In addition, the DB seeder util was moved to its own file, and updates
were made to the utility to allow it to be configured through the
application.properties file. Configurations include:
* enabling / disabling the seeder
* the number of members to generate
* the number of visits to generate