#2 Convert Membership Hierarchy To Abstract Classes
Convert Membership Hierarchy to Abstract Classes instead of interface
to make using hibernate easier. Moving forward, the abstract super class
should be treated as if it were an interface, except for the
associations required by hibernate.

Update H2 properties so that the db memory database can be queried for
debugging purposes.