Create basic cloud build configuration
This commit starts the effort to set up a continuous delivery pipeline
for deploying the application. Google's cloud builder tools will be used
as our continus delivery tools.

This commit creates a basic Dockerfile and cloudbuild.yml configuration.
At the moment, the cloud builder configuration only builds the docker
container. The build does not even push the container to a registry yet.
Eventually, it should auto deploy the new version of the
container into a kuberneties cluster.