Create deployment and service kubernetes resources
Created the deployment.yml and service.yml kubernetes resources for
deploying the backend service to a kubernetes cluster. The deployment
is as simple as:

```
kubectl create -f ci/deployment.yml
```

Assuming the resource does not yet exist. I still have to figure out how
to patch existing resources so that a build can update the deployment
rather than forcing it to be deleted and recreated.