Add Pagination to the get members endpoint
Change get members endpoint to use the jpa entity manager to dynamically
construct a query. The query has a page-offset and a page size. These
will allow the frontend to paginate the query request.

Junit tests were updated so the method signature lines up, but the tests
themselves don't convey much and probably should be re-written.