Update Gradle to Include SonarQube
Add sonarqube to the gradle build. To publish to sonarqube, the command
`gradle sonarqube` just needs to be invoked. (In order for it to run,
the gradle.properties file needs to be set up with the sonarqube
endpoint and the access key)

The cloudbuild.yml was also updated to run the sonarqube. The goal is to
eventually have the google cloud build run the builds automatilcy and
publish the sonarqube results.